+++ app/src/main/java/de/erichambuch/biproclient/AppInfo.java
@@ -2,7 +2,7 @@ package de.erichambuch.biproclient;
 
 /**
  * Autor: Eric Hambuch (erichambuch@googlemail.com)
- * Lizenz: Proprietär
+ * Lizenz: Apache License 2.0 (https://www.apache.org/licenses/LICENSE-2.0)
  * <p>
  *     Versionen: 1.0 (07.03.2021) erstes Release
  *     1.0.1 (11) (13.03.2021) - Bugfix für Scrollbar
@@ -14,6 +14,7 @@ package de.erichambuch.biproclient;
  *     1.4.0 (17) - Update Android 12
  *     1.4.1 (18) - Anzeige OSS Lizenzen
  *     1.4.2 (19) - Update Libs, Bugfixes
+ *     1.5.0 (20) - Verbessertes Logging bei Anmeldung, Android 8 als Minimum
  * </p>
  * <p>
  *     Offene Themen:<ol>