+++ app/src/main/java/de/erichambuch/biproclient/main/SettingsFragment.java
@@ -3,7 +3,6 @@ package de.erichambuch.biproclient.main;
 import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
-import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.provider.DocumentsContract;
@@ -121,9 +120,7 @@ public class SettingsFragment extends PreferenceFragmentCompat {
         Intent intent = new Intent(Intent.ACTION_OPEN_DOCUMENT);
         intent.addCategory(Intent.CATEGORY_OPENABLE);
         intent.setType("*/*"); // fester MIME-Type funktioniert nicht beim Download von Dateien
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, Uri.fromFile(Environment.getDownloadCacheDirectory()));
-        }
+        intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, Uri.fromFile(Environment.getDownloadCacheDirectory()));
         startActivityForResult(intent, REQUEST_CODE_LOAD);
     }
 
@@ -131,9 +128,7 @@ public class SettingsFragment extends PreferenceFragmentCompat {
         Intent intent = new Intent(Intent.ACTION_CREATE_DOCUMENT);
         intent.setType("application/json");
         intent.putExtra(Intent.EXTRA_TITLE, "biproconfig.json");
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, Uri.fromFile(Environment.getDownloadCacheDirectory()));
-        }
+        intent.putExtra(DocumentsContract.EXTRA_INITIAL_URI, Uri.fromFile(Environment.getDownloadCacheDirectory()));
         intent.addCategory(Intent.CATEGORY_OPENABLE);
         startActivityForResult(intent, REQUEST_CODE_SAVE);
     }